Static rewrite for remollSearchPath.hh
